added 102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
 
 int main(void)
 {
-	int i, a, b, result;
+	long int i, a, b, result;
 
 	a = 1;
 	b = 2;
